Support multiple JBoss standalone instances with same home directory
JBoss standalone mode allows multiple instances to run using same
Jboss home dir. To support this Key and Diskpath is set to the
complete configuration file used for the instance.

Also same configuration file can be used for different instances
by giving diffrenet port offset. The code is updated to check
for port offset in command line parameters. If present it will
be appended to Key and Diskpath to uniquely identify the instance.

JBoss does not mandate that standalone should be a directory in
configuration path. Updated the agent to pass detected deployment type
from appserver enumeration to JBoss instance. Value of deployment type
will be used to determine if it is an JBoss standalone instance.